+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.1.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-cli@0.1.1...@carbon/ibm-cloud-cognitive-cli@0.1.2) (2020-11-30)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive-cli
+
+
+
+
+
+## [0.1.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/tree/master/packages/cli/compare/@carbon/ibm-cloud-cognitive-cli@0.1.0...@carbon/ibm-cloud-cognitive-cli@0.1.1) (2020-11-26)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive-cli
+
+
+
+
+
+# Change Log
+
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.